refactor: split transform.rs into smaller files
